+++ app/src/main/java/com/paulsavchenko/dotsandcharts/domain/dto/PointDto.kt
@@ -1,6 +1,6 @@
 package com.paulsavchenko.dotsandcharts.domain.dto
 
 data class PointDto(
-    val pointX: Float,
-    val pointY: Float,
+    val pointX: Double,
+    val pointY: Double,
 )
